+++ src/models/InputModel.cs
@@ -1,21 +1,41 @@
 // Copyright (C) 2023 Karl Pickett / ViLark Project
+using System.Diagnostics;
 namespace vilark;
+
+// This can be a file, or a buffer, or a simple string choice
+interface ISelectableItem
+{
+    // This is what the user sees, e.g. "github/myproject | foo.txt"
+    public string GetDisplayString();
+
+    // This is what the external program receives, e.g. /home/bob/github/myproject/foo.txt
+    // Note that for files, we always return absolute paths
+    public string GetChoiceString();
+
+    // This is what the user searches on, which is a subset of the path.
+    // example: github/myproject/foo.txt
+    public string GetSearchString();
+}
+
+
 
 
 class InputModel
 {
-    private IEnumerable<IScrollItem>? _data = null;
-    private IEnumerable<IScrollItem>? _filtered_data = null;
+    private IEnumerable<ISelectableItem>? _data = null;
+    private IEnumerable<ISelectableItem>? _filtered_data = null;
     private Thread? _load_thread = null;
     private OptionsModel m_options;
-    private EventQueue<Notification> m_load_event;
+    private Config m_config;
+    private EventQueue<Notification> m_notifications;
 
     // This is null when we are still loading
-    public IEnumerable<IScrollItem>? FilteredData => _filtered_data;
+    public IEnumerable<ISelectableItem>? FilteredData => _filtered_data;
 
-    public InputModel(OptionsModel options, EventQueue<Notification> loadEvent) {
+    public InputModel(OptionsModel options, Config config, EventQueue<Notification> notifications) {
+        m_config = config;
         m_options = options;
-        m_load_event = loadEvent;
+        m_notifications = notifications;
     }
 
     public void StartLoadingAsync() {
@@ -28,14 +48,18 @@ class InputModel
 
     private void LoadInput(string selectedDirectory) {
         try {
+            var stopwatch = new Stopwatch();
+            stopwatch.Start();
             var entries = LoadInputImpl(selectedDirectory);
-            m_load_event.AddEvent(new Notification(CompletedData: entries));
+            stopwatch.Stop();
+            Log.Info($"LoadInputImpl completed in {stopwatch.Elapsed}");
+            m_notifications.AddEvent(new Notification(CompletedData: entries));
         } catch (Exception e) {
-            m_load_event.AddEvent(new Notification(ErrorMessage: e.ToString()));
+            m_notifications.AddEvent(new Notification(ErrorMessage: e.ToString()));
         }
     }
 
-    private IEnumerable<IScrollItem> LoadInputImpl(string selectedDirectory) {
+    private IEnumerable<ISelectableItem> LoadInputImpl(string selectedDirectory) {
         var inputFileName = Environment.GetEnvironmentVariable("VILARK_INPUT_FILE");
         if (inputFileName != null) {
             // An explicit input file was given.
@@ -49,19 +73,19 @@ class InputModel
                     if (line == null) {
                         break;
                     }
-                    entries.Add(new ExternalInputEntry {itemName=line, displayAsFile=displayAsFile});
+                    entries.Add(new ExternalInputEntry(line, displayAsFile:displayAsFile));
                 }
             }
             return entries;
         } else {
             // Load files, recursively, and honor all ignore rules
-            var explorer = new DirectoryExplorer(selectedDirectory, m_load_event);
+            var explorer = new DirectoryExplorer(m_config, selectedDirectory, m_notifications);
             var entries = explorer.Scan();
             return entries;
         }
     }
 
-    public void SetCompletedData(IEnumerable<IScrollItem> data) {
+    public void SetCompletedData(IEnumerable<ISelectableItem> data) {
         _data = data;
     }
 
@@ -70,10 +94,10 @@ class InputModel
             return;
         }
         if (searchText != String.Empty) {
-            List<IScrollItem> filtered = new();
+            List<ISelectableItem> filtered = new();
             var query = new FuzzyTextQuery(searchText, mode);
             foreach (var dentry in _data) {
-                var result = query.RunQuery(dentry.GetSelectionString());
+                var result = query.RunQuery(dentry.GetSearchString());
                 if (result.is_complete_match) {
                     filtered.Add(dentry);
                 }